# backend/api/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: Database API for Nefryt LDS
  title: Nefryt LDS API
  version: 1.0.0
servers:
- description: Test server
  url: http://192.168.1.231:8080/
- description: Developer's local server
  url: http://localhost:8080/
security:
- bearerAuth: []
paths:
  /auth/login:
    post:
      operationId: auth_login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginPermissions'
          description: OK
      security: []
      summary: Authentication Login
      tags:
      - auth
      x-openapi-router-controller: api.controllers.auth_controller
  /auth/refresh:
    post:
      operationId: auth_refresh
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginPermissions'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Authentication token refresh
      tags:
      - auth
      x-openapi-router-controller: api.controllers.auth_controller
  /event:
    get:
      description: List all events
      operationId: list_events
      parameters:
        - description: Page number
          in: query
          name: $page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - description: Page size
          in: query
          name: $size
          required: false
          schema:
            type: integer
            maximum: 10000
            minimum: 1
            default: 50
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Event'
                type: array
          description: A array of events
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List events
      tags:
      - events
      x-openapi-router-controller: api.controllers.events_controller
  /event/{eventId}:
    get:
      description: Info for specific event
      operationId: get_event_by_id
      parameters:
      - description: The id of the event to retrieve
        explode: false
        in: path
        name: eventId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Gets detailed event
      tags:
      - events
      x-openapi-router-controller: api.controllers.events_controller
  /event/{eventId}/ack:
    post:
      description: Acknowledges an event
      operationId: ack_event
      parameters:
      - description: The id of the event to retrieve
        explode: false
        in: path
        name: eventId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Information'
          description: Updated
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Acknowledges ack
      tags:
      - events
      x-openapi-router-controller: api.controllers.events_controller
  /event_def:
    get:
      description: List all event defs
      operationId: list_event_defs
      parameters:
        - description: Page number
          in: query
          name: $page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - description: Page size
          in: query
          name: $size
          required: false
          schema:
            type: integer
            maximum: 10000
            minimum: 1
            default: 50
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventDef'
                type: array
          description: An array of event defs
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List event defs
      tags:
        - events
    post:
      description: Create an event def
      operationId: create_event_def
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDef'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDef'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create event def
      tags:
        - events
  /event_def/{eventDefId}:
    delete:
      description: Delete specific event def
      operationId: delete_event_def_by_id
      parameters:
        - description: The id of the event def to retrieve
          explode: false
          in: path
          name: eventDefId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "204":
          description: No content
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Deletes event def
      tags:
        - events
    get:
      description: Info for specific event def
      operationId: get_event_def_by_id
      parameters:
        - description: The id of the event def to retrieve
          explode: false
          in: path
          name: eventDefId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDef'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Gets event def details
      tags:
        - events
    put:
      description: Update event def
      operationId: update_event_def
      parameters:
        - description: The id of the event def to retrieve
          explode: false
          in: path
          name: eventDefId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDef'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDef'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Updates event def
      tags:
        - events
...
components:
...
  schemas:
...
    Event:
      example:
        MethodID: 6
        Details: Details
        Position: 1
        ID: 0
        EventDefID: NO_DATA
        Verbosity: Verbosity
        EndDate: 2000-01-23T04:56:07.000+00:00
        Silient: true
        AckDate: 2000-01-23T04:56:07.000+00:00
        Caption: Caption
        BeginDate: 2000-01-23T04:56:07.000+00:00
      properties:
        ID:
          description: none
          example: 0
          format: int64
          readOnly: true
          title: ID
          type: integer
        EventDefID:
          description: none
          title: EventDefID
          type: string
        MethodID:
          description: none
          title: MethodID
          type: integer
        BeginDate:
          description: none
          format: date-time
          title: BeginDate
          type: string
        AckDate:
          description: none
          format: date-time
          title: AckDate
          type: string
        EndDate:
          description: none
          format: date-time
          title: EndDate
          type: string
        Details:
          description: none
          title: Details
          type: string
        Position:
          description: none
          title: Position
          type: integer
        Caption:
          description: none
          title: Caption
          type: string
        Verbosity:
          description: none
          title: Verbosity
          type: string
        Silient:
          description: none
          title: Silient
          type: boolean
      required:
      - BeginDate
      - EventDefID
      - MethodID
      title: Event
      type: object
    EventDef:
      example:
        ID: ANOMALY
        Verbosity: WARNING
        Caption: Caption
        Silent: false
        Visible: true
        Enabled: true
      properties:
        ID:
          description: none
          readOnly: true
          title: ID
          type: string
        Verbosity:
          description: none
          title: Verbosity
          type: string
        Caption:
          description: none
          title: Caption
          type: string
        Silent:
          description: none
          title: Silent
          type: boolean
        Visible:
          description: none
          title: Visible
          type: boolean
        Enabled:
          description: none
          title: Enabled
          type: boolean
      required:
        - ID
        - Name
      title: TrendDef
      type: object
...
    Error:
      properties:
        code:
          format: int32
          title: code
          type: integer
        message:
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
    Information:
      example:
        message: message
        affected: 6
        status: 0
      properties:
        status:
          format: int32
          nullable: false
          title: status
          type: integer
        affected:
          format: int32
          nullable: false
          title: affected
          type: integer
        message:
          nullable: true
          title: message
          type: string
      required:
      - affected
      - message
      - status
      title: Information
      type: object
    Login:
      example:
        password: password
        deviceId: deviceId
        deviceName: deviceName
        username: username
      properties:
        username:
          minLength: 1
          title: username
          type: string
        password:
          minLength: 1
          title: password
          type: string
        deviceId:
          minLength: 1
          title: deviceId
          type: string
        deviceName:
          minLength: 1
          title: deviceName
          type: string
      required:
      - password
      - username
      title: Login
      type: object
      x-examples:
        example-1:
          username: JohnDoe
          password: _hashed_password_here_
          deviceId: ad49afd9-ec38-4716-b93f-416bd8c0c684
          deviceName: model
    LoginPermissions:
      example:
        refreshTokenExpiration: refreshTokenExpiration
        success: true
        permissions:
        - key: '{}'
        - key: '{}'
        username: username
        token: token
        refreshToken: refreshToken
      properties:
        username:
          minLength: 1
          title: username
          type: string
        success:
          title: success
          type: boolean
        token:
          minLength: 1
          title: token
          type: string
        refreshToken:
          minLength: 1
          title: refreshToken
          type: string
        refreshTokenExpiration:
          minLength: 1
          format: date-time
          title: refreshTokenExpiration
          type: string
        permissions:
          items:
            additionalProperties: true
            type: object
          title: permissions
          type: array
      required:
      - refreshToken
      - refreshTokenExpiration
      - token
      title: LoginPermissions
      type: object
      x-examples:
        example-1:
          username: JohnDoe
          success: true
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
          refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
          refreshTokenExpiration: 2022-03-30T23:59:59.000Z
          permissions:
          - TRENDS
          - ANALYTICS
          - ADMIN
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http